feat: Organize project structure by adding services, components, and router modules
