+++ .gitignore
@@ -12,3 +12,4 @@
 **/*.rs.bk
 
 .vscode/
+_command
