+++ Cargo.toml
@@ -1,17 +1,28 @@
 [package]
 name = "prime-number-generator"
 version = "0.1.0"
-authors = ["sho5081857"]
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
 
-dioxus = { version = "0.5", features = ["web", "router"] }
+dioxus = { version = "0.6", features = ["web", "router"] }
 
 getrandom ={ version = "0.2.3", features = ["js"] }
 rand = "0.8.5"
 
 # Debug
 dioxus-logger = "0.5.1"
+
+[profile]
+
+[profile.wasm-dev]
+inherits = "dev"
+opt-level = 1
+
+[profile.server-dev]
+inherits = "dev"
+
+[profile.android-dev]
+inherits = "dev"
