+++ src/components.rs
@@ -1,0 +1,5 @@
+pub mod blog;
+pub mod generate_random_prime;
+pub mod generate_random_primes;
+pub mod home;
+pub mod is_prime;
