+++ src/components/blog.rs
@@ -1,0 +1,11 @@
+use dioxus::prelude::*;
+
+use crate::router::Route;
+
+#[component]
+pub fn Blog(id: i32) -> Element {
+    rsx! {
+        Link { to: Route::Home {}, "Go to counter" }
+        "Blog post {id}"
+    }
+}
