+++ src/services.rs
@@ -1,0 +1,2 @@
+pub mod generate_random_prime;
+pub mod generate_random_primes;
